Remove gitignore
This file doesn't need to be standardized. Excluding `.DS_Store' is
operating system specific, and `tls-redirector' is a build artifact
that should never be checked in.

Given pull requests and patches on the mailing list, junk files can
be excluded before they reach this repo, keeping it clean and tidy.