2.1: Simplify PORT, make systemd truly optional
This commit changes the behavior of `PORT' subtly to no longer require
`PORT=systemd' for systemd socket activation. Instead, it's detected
proactively if tls-redirector is launched without a PORT variable.

Additionally, the systemd code was moved out of main into files that
have conditional build tags. If socket activation isn't desired by the
user, they can compile the program to not have it.

PORT=systemd was retained for backwards compatibility, and also serves
a dual purpose of forcing tls-redirector to use socket activation if
that is desired, i.e. for debugging.