Pull in dependencies when building container
Recent changes to break up the project and add tests has resulted in Dockerfile
no longer building a container. This commit changes the script to copy
everything so the internal package `socketactivation' is pulled in as well as
go.{mod,sum}.

Additionally, `strip' from `binutils' was pulled in so the resulting binary
is smaller.